+++ src/main/groovy/com/ullink/NuGetSources.groovy
@@ -1,6 +1,9 @@
 package com.ullink
 
 import org.gradle.api.GradleException
+import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.InputFile
+import org.gradle.api.tasks.Optional
 
 class NuGetSources extends BaseNuGet {
     
@@ -8,16 +11,32 @@ class NuGetSources extends BaseNuGet {
         add, remove, enable, disable, list, update
     }
 
-    Operation operation    
+    @Input
+    Operation operation
+    @Optional
+    @Input
     def sourceName
+    @Optional
+    @Input
     def sourceUrl
+    @Optional
+    @Input
     def username
+    @Optional
+    @Input
     def password
-    def configFile
+    @Optional
+    @InputFile
+    File configFile
+    @Input
     def storePasswordInClearText = false
 
     NuGetSources() {
         super('sources')
+    }
+
+    void setConfigFile(String path) {
+        configFile = project.file(path)
     }
 
     @Override