+++ src/main/groovy/com/ullink/NugetLocalsClear.groovy
@@ -1,7 +1,10 @@
 package com.ullink
 
+import org.gradle.api.tasks.Input
+
 
 class NugetLocalsClear  extends BaseNuGet{
+    @Input
     def all = false
 
     NugetLocalsClear() {